05_web_security/break_jwt_authentification/requirements.txt
Flask==3.1.2
PyJWT==2.10.1
requests==2.32.5
